main/requirements.txt
Flask
mysql-connector-python
mysqlclient
bcrypt
cryptography
python-dotenv
gunicorn
Pillow
resend